Rohan's Love for Matrix
Rohan has a special love for matrices especially for the first element of the matrix. Being good at Mathematics, he also loves to solve the different problems on matrices. So one day he started to multiply the matrix with the original matrix.  The elements of the original matrix a are given by [a00=1, a01=1, a10=1, a11=0].
Given the power of the matrix, n calculates the an and returns the a10 element mod 1000000007.

Example 1:

Input: 
n = 3
Output: 
2 
Explanation: Take the cube of the original matrix 
i.e a3 and the (a10)th element is 2.
Example 2:

Input: 
n = 4
Output: 
3
Explanation: Take the quartic of the original matrix 
i.e a4 and the (a10)th element is 3.
Your Task:  
You don't need to read input or print anything. Complete the function firstElement() which takes n as the input parameter and returns the a10 element mod 1000000007 of an.

Expected Time Complexity: O(n)
Expected Auxiliary Space: O(1)

Constraints:
1<= n <=10^6